+++ tests/Integration/API/UserAttributesTest.php
@@ -50,12 +50,12 @@ class UserAttributesTest extends AbstractAPITest
      */
     public function testUpdateUserAttribute()
     {
-        $response = static::$xsollaClient->UpdateUserAttribute([
+        static::$xsollaClient->UpdateUserAttribute([
             'project_id' => static::$projectId,
             'user_attribute_id' => static::$attributeId,
             'request' => $this->userAttribute,
         ]);
-        static::assertSame(204, $response->getStatusCode());
+        static::assertTrue(true);
     }
 
     /**
@@ -74,10 +74,10 @@ class UserAttributesTest extends AbstractAPITest
      */
     public function testDeleteUserAttribute()
     {
-        $response = static::$xsollaClient->DeleteUserAttribute([
+        static::$xsollaClient->DeleteUserAttribute([
             'project_id' => static::$projectId,
             'user_attribute_id' => static::$attributeId,
         ]);
-        static::assertSame(204, $response->getStatusCode());
+        static::assertTrue(true);
     }
 }